DGNv8/DWG: win64 related fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@37969 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/dwg/ogr_dwg.h
@@ -190,7 +190,7 @@ class OGRDWGDataSource : public OGRDataSource
 
     const char          *GetName() override { return m_osName; }
 
-    int                 GetLayerCount() override { return apoLayers.size(); }
+    int                 GetLayerCount() override { return static_cast<int>(apoLayers.size()); }
     OGRLayer            *GetLayer( int ) override;
 
     int                 TestCapability( const char * ) override;
